Implemented main functionality. Changed funcs.

+++ src/sample/Grid.java
@@ -1,8 +1,8 @@
 package sample;
 
 public interface Grid {
-    public void setCoordinates(int x, int y);
-    public void unsetCoordinates(int x, int y);
+    public void setCoordinates(double x, double y);
+    public void unsetCoordinates(double x, double y);
     public char[][] getCompleteGrid();
     public void setCompleteGrid(char [][] grid);
 }
